requirements.txt
mujoco
loguru
pyyaml
scipy
torch
onnxruntime
pynput
ipdb
termcolor
sshkeyboard
pygame
zmq
joblib
pyvicon-datastream
huggingface-hub
# unitree_sdk2py
# git+https://github.com/unitree-ai/unitree_sdk2py.git